refactor: restructuring
