+++ src/main/java/zone/vao/nexoAddon/events/nexo/blocks/NexoBlockBreakListener.java
@@ -4,10 +4,10 @@ import com.nexomc.nexo.api.events.custom_block.chorusblock.NexoChorusBlockBreakE
 import com.nexomc.nexo.api.events.custom_block.noteblock.NexoNoteBlockBreakEvent;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import zone.vao.nexoAddon.events.nexo.blocks.playerNexoBlockBreaks.DropExperienceListener;
-import zone.vao.nexoAddon.events.nexo.blocks.playerNexoBlockBreaks.InfestedListener;
-import zone.vao.nexoAddon.events.nexo.blocks.playerNexoBlockBreaks.MiningToolsListener;
-import zone.vao.nexoAddon.events.nexo.blocks.playerNexoBlockBreaks.ShiftBlockListener;
+import zone.vao.nexoAddon.events.nexo.blocks.breaks.DropExperienceListener;
+import zone.vao.nexoAddon.events.nexo.blocks.breaks.InfestedListener;
+import zone.vao.nexoAddon.events.nexo.blocks.breaks.MiningToolsListener;
+import zone.vao.nexoAddon.events.nexo.blocks.breaks.ShiftBlockListener;
 import zone.vao.nexoAddon.utils.BlockUtil;
 
 public class NexoBlockBreakListener implements Listener {
