+++ src/main/java/zone/vao/nexoAddon/events/nexo/blocks/NexoBlockInteractListener.java
@@ -3,7 +3,7 @@ package zone.vao.nexoAddon.events.nexo.blocks;
 import com.nexomc.nexo.api.events.custom_block.noteblock.NexoNoteBlockInteractEvent;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import zone.vao.nexoAddon.events.nexo.blocks.nexoBlocksInteracts.ShiftBlockListener;
+import zone.vao.nexoAddon.events.nexo.blocks.interacts.ShiftBlockListener;
 
 public class NexoBlockInteractListener implements Listener {
